refactor(car): move RegisterCarDTO and UpdateCarDTO to car/dtos
