+++ src/modules/car/useCases/registerCar/RegisterCarController.ts
@@ -3,7 +3,7 @@ import { container } from 'tsyringe';
 import { AppError } from '@shared/errors/AppError';
 import { dataValidation } from '@shared/utils/dataValidation';
 import { RegisterCarUseCase } from './RegisterCarUseCase';
-import { RegisterCarDTO } from './RegisterCarDTO';
+import { RegisterCarDTO } from '../../dtos/RegisterCarDTO';
 
 export class RegisterCarController {
   async handle(request: Request, response: Response): Promise<Response> {
